+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Next Version
+
+### Features:
+
+* Insert badge with the number of commits of the current pull request next to the "Commits" tab,
+  closes [issue #163](https://github.com/refined-bitbucket/refined-bitbucket/issues/163),
+  [pull request #164](https://github.com/refined-bitbucket/refined-bitbucket/pull/164).
+
 # 3.7.3 (2018-03-08)
 
 From now on, you have the option to disable update notifications that open a new browser tab each time a new version